Refactor Cedrik to an OTP application
This is a major change that breaks backward compatibility.

Before this commit Cedrik was a library, a collection of modules and
functions. After this commit Cedrik is an OTP application and its
internal structure is more idiomatic Elixir (ie GenServers and
Supervisors).

Hopefully this will make it easier to create a demo UI application
as planned.